Handle client disconnection by DRS_process_comm_fds.

+++ TcpClient.h
@@ -10,7 +10,7 @@ typedef struct TcpClient TcpClient;
 #include "TcpServerController.h"
 
 typedef struct TcpClient {
-    int comm_fd;
+    uintptr_t comm_fd;
     uint32_t ip_addr;
     uint16_t port_no;
     TcpServerController *tsc;
